fix: 캐릭터 애니메이션 재생을 위한 delta time 계산 방식 변경

+++ src/public/Animation.h
@@ -3,13 +3,16 @@
 
 class Animation {
 public:
-    Animation() : mTimer(0), mFrameCount(0) {}
+    Animation() : mTimer(0), mFrameCount(0) {
+    }
+
     Animation(int frameCount, float length);
+
     ~Animation();
 
     void step(const float deltaTime) { mTimer.step(deltaTime); }
     float getLength() const { return mTimer.getLength(); }
-    int getCurrentFrame() const { return static_cast<int>(mTimer.getElapsedTime() / mTimer.getLength()) * mFrameCount; }
+    int getCurrentFrame() const { return static_cast<int>(mTimer.getElapsedTime() / mTimer.getLength() * static_cast<float>(mFrameCount)); }
 
 private:
     Timer mTimer;
